backend/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
authors = [{name = "halilagin", email = "halil.agin@gmail.com"}]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi>=0.115.11",
    "uvicorn[standard]==0.24.0",
    "sqlalchemy==2.0.23",
    "pydantic[email]>=2.10.6",
    "python-jose[cryptography]==3.3.0",
    "passlib>=1.7.4",
    "python-multipart==0.0.6",
    "python-dotenv==1.0.0",
    "pydantic-settings==2.1.0",
    "bcrypt==3.2.2",
    "pytest>=8.3.5",
    "httpx>=0.28.1",
    "starlette>=0.46.0",
    "pytest-asyncio>=0.25.3",
    "reportlab>=4.3.1",
    "pypdf2>=3.0.1",
    "pypdf>=5.3.1",
    "endesive>=2.18.3",
    "pyjwt>=2.10.1",
    "stripe>=11.6.0",
    "psycopg2-binary>=2.9.10",
    "pycryptodome>=3.22.0",
    "jinja2>=3.1.4",
    "pillow>=11.2.1",
    "aiosmtplib>=3.0.1",
    "apscheduler>=3.10.4",
    "pytz>=2025.2",
    "openai>=1.98.0",
    "pydub>=0.25.1",
    "opencv-python>=4.12.0.88",
    "ray[serve]>=2.48.0",
    "grpcio==1.62.2",
    "aiofiles>=24.1.0",
    "setuptools>=80.9.0",
    "psutil>=7.0.0",
    "requests>=2.31.0",
    "ffmpeg-python>=0.2.0",
    "pysubs2>=1.8.0",
    "boto3>=1.40.45",
    "redis>=7.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
